Extract onDataViewChange() from Applab
